third_party/blink/web_tests/fast/forms/placeholder-stripped-expected.txt

CR and LF in placeholder values should be stripped for input type. For textarea should not be stripped.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.visiblePlaceholder(input0) is "first line second line"
PASS internals.visiblePlaceholder(input1) is ""
PASS internals.visiblePlaceholder(textarea0) is "first line \nsecond line"
PASS internals.visiblePlaceholder(textarea1) is ""
PASS internals.visiblePlaceholder(textarea2) is ""
PASS successfullyParsed is true

TEST COMPLETE
